[BoundsWidening] Update observed bounds with widened bounds
For each statement update the observed bounds in the BlockState with the
widened bounds as computed by the updated bounds widening analysis. The bounds
of variables whose bounds are killed are reset to the declared bounds. This PR
also updates the method GetBoundsWidenedAndNotKilled that returns the bounds
widened in a block before a given statement and not killed by that statement.

This PR also deletes the old bounds widening analysis files
BoundsAnalysis(.h,.cpp).
