+++ clang/lib/Sema/CMakeLists.txt
@@ -23,7 +23,6 @@ clang_tablegen(OpenCLBuiltins.inc -gen-clang-opencl-builtins
 add_clang_library(clangSema
   AnalysisBasedWarnings.cpp
   AvailableFactsAnalysis.cpp
-  BoundsAnalysis.cpp
   BoundsWideningAnalysis.cpp
   CheckedCAlias.cpp
   CheckedCAnalysesPrepass.cpp
